Initial handle_keys functionality
Keybindings for WASD and QZ and EC should be implemented, though I'm not entirely sure what the subtle differences between them are or if they're truly implemented the way the should be. Currently they seem to be moving the look at point. I'm not sure if this is intentional. Will have to clarify and update if necessary.