imGuizmo tweaks
The rotation is now based on a delta rotation, which stops the object from constantly spinning. The rotation of the gizmo and the object are still a bit out of sinc, Like, some directions seem to track (say, left is left, right is right) while others are opposites (say up is down and down is up). Not sure what the cause is. 

Still need to update the guizmo based on the rotation of the camera object so it tracks properly. I think we need to make it so the guizmo first updates itself to match the object orientation, and then if the guizmo is moved within the widget, that'll apply the rotation to the camera.