ImGuizmo tweaks: Texture flipped
The texture of the image seemed to be flipped, as in, we were looking at it from the back. It now displays from the front and the Guizmo movement tracks in xyz with the camera. Note that now the trackball and other movements are slightly off so they'll need to be fixed. 

Next up is making sure the guizmo rotation is adjusted anytime the camera moves.