Debugging guizmo
Debugging so far seems to confirm that the matrix_to_quat function does successfully convert one mat4 matrix into a quat, which can be cast back to a mat4 and back to quat without any changes to the values.

I'm assuming the bug seen in the guizmo behavior is due to the rotate_guizmo function, specifically the matrix math that is used to set the rotation of the guizmo.