Transfer Function work
I've practiced drawing rectangles and explored the differences between a popup utility and trying to get a separate window to render. For now I feel comfortable drawing a gradient based on the color map, though I'm still not super clear on what the my_bar and color_ptr are for.